Typescript 5: Fixes #3 and #5 (#6)
fix: #5 hostOrText.readFile

In the traverse-filesystem, we use now the real
ts.sys for the resolution of filenames.

In the past, we created a proxy ts.sys that
only implemented readFile of ts.sys and was
a hacky approach.

If the virtualFs is enabled though,
we still go with the "hacked
version". This happens only in the
tests though.