refactor(core): make barrel file optional in module type
This change is a prerequisite for the upcoming barrel-less mode. The module type
has been updated so that a barrel file is no longer mandatory.

This affects the `path` property, which previously always pointed to the barrel file.
The new implementation supports both barrel-based and barrel-less (inactive) setups

Additionally, the barrel file name is added to the config but remains internal.