feat(core): rename config property `tagging` to `modules`
With barrel-less modules, modules are not defined by the existence of a barrel file
but by the `config` itself. The term `tagging` doesn't reflect the new context
anymore, as it doesn't just tag but also specify modules.

Therefore, `tagging` is deprecated and still works, but for new configs, `modules`
should be used instead.