feat: exclude `sheriff.config.ts` from sheriff eslint rules (#167)
The exclusion of the `sheriff.config.ts` is a necessary step 
because we want to separate verification processes running.

One for the `sheriff.config.ts` where common config errors are shown - 
and only there. 

The second is the actual verification of the application code, namely 
module boundaries and dependency rules.